Why:
The CI/CD jobs for web would only build for one flavor of the app. We
needed them to build, package, publish and deploy for both dari and
pashto flavors of the web applications.

This change addresses the need by:

Restructuring the overall pipeline and prerequisite environment
variables to support pipeline jobs that would factor in dari and pashto
flavors of the app.

Adjusting the jobs for 'buildWebApps', 'buildPackageBackendHost',
'containerizeWeb', and 'deployWeb' stages of the pipeline to run for
both flavors of the application. Adding in configuration and then
building one web-app at a time so that release output files can be
created for both.