Why:
Missing build job for android build stage in CI/CD pipeline

This change addresses the need by:

Adding in a job that injects configurations and runs builds for dari and
pashto versions of the android app one at a time. Making sure to clean
build outputs after each build so as not to mix configs between the two
versions.