Why:
Operations team would need to wait too long after back-end deployment to be able to test android app.

This change addresses the need by:

Having the android build process run in parallel to the build, package, and deployment of the web infrastructure by placing it in the first stage of the pipeline.